+++ forui_hooks/lib/forui_hooks.dart
@@ -10,7 +10,6 @@ export '/src/resizable_controller_hook.dart';
 export '/src/select_controller_hook.dart';
 export '/src/select_group_controller_hook.dart';
 export '/src/select_menu_tile_controller_hook.dart';
-export '/src/select_tile_group_controller_hook.dart';
 export '/src/slider_controller_hook.dart';
 export '/src/tab_controller_hook.dart';
 export '/src/time_field_controller_hook.dart';