+++ samples/lib/widgets/tile_group.dart
@@ -138,8 +138,7 @@ class _MergeTileGroupPageState extends StatefulSampleState<MergeTileGroup> {
         ],
       ),
       .selectGroup(
-        selectController: _controller,
-        children: const [
+        control: .managed(controller: _controller), children: const [
           FSelectTile(title: Text('List View'), value: 'List'),
           FSelectTile(title: Text('Grid View'), value: 'Grid'),
         ],